@@ -3,35 +3,35 @@
   "version": "0.7.1",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.7",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.15.20",
     "@mui/lab": "^5.0.0-alpha.60",
-    "@mui/material": "^5.14.7",
-    "@mui/x-date-pickers": "^5.0.0-beta.2",
-    "@sentry/react": "^8.11.0",
-    "@sentry/vite-plugin": "^2.19.0",
+    "@mui/material": "^5.15.20",
+    "@mui/x-date-pickers": "^5.0.20",
+    "@sentry/react": "^8.12.0",
+    "@sentry/vite-plugin": "^2.20.1",
     "buffer": "^6.0.3",
     "css-what": "^5.1.0",
-    "date-fns": "^2.28.0",
+    "date-fns": "^2.30.0",
     "dotenv": "^10.0.0",
-    "firebase": "^9.23.0",
-    "firebase-tools": "^11.30.0",
+    "firebase": "^10.12.2",
+    "firebase-tools": "^13.11.4",
     "glob-parent": "^6.0.2",
-    "jose": "^4.9.2",
-    "react": "^18.2.0",
+    "jose": "^5.4.1",
+    "react": "^18.3.1",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.3.1",
-    "react-router-dom": "^6.0.2",
+    "react-dom": "^18.3.1",
+    "react-icons": "^4.12.0",
+    "react-router-dom": "^6.24.0",
     "react-uuid": "^2.0.0",
     "rollup": "^4.18.0",
-    "vite": "^5.2.13",
+    "vite": "^5.3.1",
     "web-vitals": "^1.1.2"
   },
   "devDependencies": {
-    "@babel/plugin-transform-react-jsx": "^7.16.7",
+    "@babel/plugin-transform-react-jsx": "^7.24.7",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-json": "^6.1.0",
@@ -41,8 +41,6 @@
     "@vitejs/plugin-react": "^4.3.1",
     "@vitest/coverage-v8": "^1.6.0",
     "jsdom": "^24.1.0",
-    "rollup": "^4.18.0",
-    "vite": "^5.2.13",
     "vitest": "^1.6.0"
   },
   "scripts": {
